worker/requirements.txt
celery
redis
agentscope
jsonschema
hiredis
sqlalchemy
python-dotenv
requests
duckduckgo-search
ollama
yfinance
